+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 0.13.2
 - Fix for "Fix including nested directories".
 - Step history.
+- A lot of refactoring regarding history/reverse debugging.
 
 # 0.13.1
 - Code coverage for internal simulator.